WTA #31: Moved HomePageNetworkData to separate file and moved common error handling logic.
