+++ core/models/src/main/java/com/jacob/wakatimeapp/core/models/Error.kt
@@ -14,7 +14,7 @@ sealed class Error : Exception() {
         abstract val statusCode: Int
 
         data class NoConnection(override val message: String) : NetworkErrors() {
-            override val statusCode = SERVICE_UNAVAILABLE
+            override val statusCode = -1
         }
 
         data class GenericError(override val message: String) : NetworkErrors() {
@@ -32,15 +32,12 @@ sealed class Error : Exception() {
             fun create(message: String, code: Int? = null): NetworkErrors = when (code) {
                 null -> GenericError(message)
                 in 400..499 -> ClientError(message, code)
-                SERVICE_UNAVAILABLE -> NoConnection(message)
                 in 500..599 -> ServerError(message, code)
                 else -> {
                     Timber.e("Unknown error code: $code, message: $message")
                     TODO("Handle this case")
                 }
             }
-
-            private const val SERVICE_UNAVAILABLE = 503
         }
     }
 }
